update deplyoment script

+++ src/AbstractBlocklockReceiver.sol
@@ -13,9 +13,9 @@ import {ConfirmedOwner} from "./access/ConfirmedOwner.sol";
 abstract contract AbstractBlocklockReceiver is IBlocklockReceiver, ConfirmedOwner {
     /// @notice BlocklockSender contract for conditional encryption requests, subscription management and handling decryption keys
     IBlocklockSender public blocklock;
-    
+
     /// @notice Event to log direct transfer of native tokens to the contract
-    event Received(address, uint);
+    event Received(address, uint256);
 
     /// @notice Event to log deposits of native tokens
     event Funded(address indexed sender, uint256 amount);
@@ -188,7 +188,7 @@ abstract contract AbstractBlocklockReceiver is IBlocklockReceiver, ConfirmedOwne
         blocklock.cancelSubscription(subscriptionId, to);
     }
 
-    /// @notice The receive function is executed on a call to the contract with empty calldata. 
+    /// @notice The receive function is executed on a call to the contract with empty calldata.
     /// This is the function that is executed on plain Ether transfers (e.g. via .send() or .transfer()).
     receive() external payable {
         emit Received(msg.sender, msg.value);
